refactor(update): split UpdateReminder/hooks.tsx into AppUpdate/* + 165-test suite
Move the 1288-line UpdateReminder/hooks.tsx grab-bag into a dedicated
packages/kit/src/components/AppUpdate/ module split by concern, and
extend the migrated test file to cover the new pure-function modules
that the original test suite did not reach.

Source split:
- updateMutex.ts          JS-process-wide single-flight for downloadPackage
- updateAnalytics.ts      Mixpanel param builder + attempt-id lifecycle
- updateRetry.ts          Exponential backoff + isUnrecoverable bail
- updateErrorTaxonomy.ts  PII-scrubbing + low-cardinality error code map
- updateDialogs.tsx       UPDATE_DIALOG_INTERVAL throttled dialog UI
- AppUpdateForeground.tsx Mount-once foreground container + once-per-app guard
- useAppUpdate.tsx        useDownloadPackage + useAppUpdateInfo + re-exports
- index.ts                Public surface

Test coverage (165 tests across 4 suites, all passing):
- useAppUpdate.test.ts: migrated from UpdateReminder/hooks.test.ts.
  Imports rewired to ./useAppUpdate. Added: 2 Android path-redaction
  cases, 19 isUnrecoverableDownloadError cases, 2 once-per-app-lifetime
  guard cases, 5 getSkipGPGVerification AND-chain cases (the
  security-sensitive bundle signature bypass — verifies that both the
  native isSkipGpgVerificationAllowed gate AND the dev-setting toggle
  must be true, and that a thrown native gate fails closed).
- updateMutex.test.ts:    8 new — single-flight collapse, slot cleanup
                          on resolve/reject, concurrent-rejection share
- updateAnalytics.test.ts: 27 new — exhaustive platform branch table,
                          attempt-id lazy-init/rotate/idempotence,
                          unknown-strategy fallback, null-version safety
- updateDialogs.test.tsx: 10 new — 24h throttle + boundary cases
                          (=, <, > and falsy lastShownAt)
